Fix benchmark tests
Ensure that we're comparing only call performance and not compiler performance. This requires that Logtalk message sending goals are actually compiled when the source files are compiled and not postponed to runtime. The fix have two parts: (1) calls the benchmark goals from within a Logtalk object so that the Logtalk compiler gets a chance to compile the message sending calls and (2) adds missing meta-predicate directives so that the argument of the benchmark predicates is declared as a meta-argument, a goal in this case, and thus also compiled.